requirements.txt
git+https://github.com/matteobettini/benchmarl_sphinx_theme.git
torchrl>=0.2.0
torch
tqdm
hydra-core
vmas>=1.2.10
eztils==0.4.94
wandb==0.16.6
numpy==1.26.4
omegaconf==2.3.0



